+++ sdk/evaluation/azure-ai-evaluation/azure/ai/evaluation/__init__.py
@@ -21,7 +21,7 @@ from ._evaluators._intent_resolution import IntentResolutionEvaluator
 from ._evaluators._meteor import MeteorScoreEvaluator
 from ._evaluators._protected_material import ProtectedMaterialEvaluator
 from ._evaluators._qa import QAEvaluator
-from ._evaluators._completeness import CompletenessEvaluator
+from ._evaluators._response_completeness import ResponseCompletenessEvaluator
 from ._evaluators._task_adherence import TaskAdherenceEvaluator
 from ._evaluators._relevance import RelevanceEvaluator
 from ._evaluators._retrieval import RetrievalEvaluator
@@ -58,7 +58,7 @@ __all__ = [
     "FluencyEvaluator",
     "GroundednessEvaluator",
     "GroundednessProEvaluator",
-    "CompletenessEvaluator",
+    "ResponseCompletenessEvaluator",
     "TaskAdherenceEvaluator",
     "IntentResolutionEvaluator",
     "RelevanceEvaluator",