+++ sdk/evaluation/azure-ai-evaluation/azure/ai/evaluation/_common/constants.py
@@ -6,7 +6,7 @@ from enum import Enum
 from azure.core import CaseInsensitiveEnumMeta
 
 PROMPT_BASED_REASON_EVALUATORS = ["coherence", "relevance", "retrieval", "groundedness", "fluency", "intent_resolution",
-                                  "tool_call_accurate", "completeness", "task_adherence"]
+                                  "tool_call_accurate", "response_completeness", "task_adherence"]
 
 
 class CommonConstants: